Updated the I2C error callsite parameter from uint32_t to uint16_t; changed the handling of I2C errors in the UART module to do either the standard error reporting or use the global error message.

+++ i2cBridge.cydsn/i2cGen2.h
@@ -72,7 +72,7 @@
     
     /// Definition of the error callback function that should be invoked when
     /// an error occurs.
-    typedef void (*I2cGen2_ErrorCallback)(I2cGen2Status, uint32_t);
+    typedef void (*I2cGen2_ErrorCallback)(I2cGen2Status, uint16_t);
     
     
     // === FUNCTIONS ===========================================================
